Add files via upload
Greetings, this is project two, a form validation program written in JavaScript.
This form doesn't get sent anywhere. Rather, this is an exercise of client side validation.
This could be used to validate a form before you send it to a server, where it will be further evaluated.

Have a great day!